+++ apm-sniffer/apm-sdk-plugin/graphql-plugin/graphql-12.x-15.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/graphql/v12_v15/GraphqlInterceptor.java
@@ -16,7 +16,7 @@
  *
  */
 
-package org.apache.skywalking.apm.plugin.graphql.v12;
+package org.apache.skywalking.apm.plugin.graphql.v12_v15;
 
 import graphql.execution.ExecutionPath;
 import graphql.execution.ExecutionStrategyParameters;
