+++ apm-sniffer/apm-sdk-plugin/graphql-plugin/graphql-12.x-15.x-plugin/src/main/resources/skywalking-plugin.def
@@ -14,4 +14,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-graphql=org.apache.skywalking.apm.plugin.graphql.v12.define.GraphqlInstrumentation
+graphql=org.apache.skywalking.apm.plugin.graphql.v12_v15.define.GraphqlInstrumentation
